parse-options: squelch usage help on certain errors
"git branch --contains no-such-commit" gives a full usage help after
showing a more relevant error message.

Usually, the usage should be shown only if the user does not know
what options are available. If the user specifies an invalid value,
the user is already aware of the available options. In this case,
there is no point in displaying the usage anymore.

Teach parse-options that certain errors during option parsing should
not give usage help, and use that throughout the system.

Signed-off-by: Paul-Sebastian Ungureanu <ungureanupaulsebastian@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>